+++ TriviaApp/TriviaAppTests/QuizUIIntegrationTests.swift
@@ -67,61 +67,4 @@ class QuizUIIntegrationTests: XCTestCase {
     private func makeQuestion(title: String = "", answers: [Answer] = []) -> Question {
         Question(id: UUID(), title: title, answers: answers, correctAnswer: Answer(id: UUID(), text: ""))
     }
-
-    private class ExaminerSpy: ExaminerDelegate {
-        var startCallCount = 0
-        var startResult: Question?
-
-        func start() throws -> Question {
-            startCallCount += 1
-
-            guard let startResult = startResult else {
-                throw NSError(domain: "test error", code: 999)
-            }
-
-            return startResult
-        }
-
-        func respond(_ question: Question, with answer: Answer) -> Question? {
-            nil
-        }
-
-        func evaluate() -> Score {
-            Score(points: 0, responses: [])
-        }
-
-        func completeLoadWithError() {
-            startResult = nil
-        }
-
-        func completeLoadWithSuccess(question: Question) {
-            startResult = question
-        }
-    }
-}
-
-extension QuizViewController {
-    var isShowingStartRetry: Bool {
-        guard let retryButton = toolbarItems?[1] else { return false }
-        return retryButton.title == "Retry"
-    }
-
-    var questionTitle: String? {
-        questionTitleLabel.text
-    }
-
-    var canSubmit: Bool {
-        guard let submitButton = toolbarItems?[1] else { return false }
-        return submitButton.isEnabled
-    }
-
-    func simulateTapOnRetry() {
-        guard let retryButton = toolbarItems?[1] else { return }
-        _ = retryButton.target?.perform(retryButton.action)
-    }
-
-    func simulateOptionIsVisible(at index: Int) -> UITableViewCell? {
-        let indexPath = IndexPath(row: index, section: 0)
-        return tableView.dataSource?.tableView(tableView, cellForRowAt: indexPath)
-    }
 }
